requirements.txt
openai==1.46.0
google-generativeai==0.8.1
anthropic==0.34.2
mistralai==0.4.2
numpy==2.1.1
termcolor==2.4.0
natsort==8.4.0
hydra-core==1.3.2
matplotlib==3.9.2
pandas==2.2.2
tqdm
